+++ src/main/java/hello/repositories/ChannelRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 
-public interface ChannelRepository extends CrudRepository<Channel, Integer> {
+public interface ChannelRepository extends CrudRepository<Channel, Long> {
     List<Channel> findAll();
 }
