+++ src/main/java/hello/repositories/MessageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 
-public interface MessageRepository extends CrudRepository<MessageOutput, Integer> {
+public interface MessageRepository extends CrudRepository<MessageOutput, Long> {
     List<MessageOutput> findAll();
 }
